sunxi: video: Support LCD and HDMI output on H6/D1
This patch series is my first attempt at supporting an LCD panel on the
T113. I can confirm this specific use case works, but I've added code
for the H6/H616 too in the process as a lot of the logic is the same.

I've also included an untested patch for HDMI support at the end in case
anyone wants to work on it. It compiles and looks like it should work,
but I haven't ungated/reset the slow HDMI clock, or HDMI sub.

To: Jagan Teki <jagan@amarulasolutions.com>
To: Andre Przywara <andre.przywara@arm.com>
To: Tom Rini <trini@konsulko.com>
To: Anatolij Gustschin <agust@denx.de>
Cc: u-boot@lists.denx.de
Signed-off-by: John Watts <contact@jookia.org>

---
Changes in v2:
- EDITME: describe what is new in this series revision.
- EDITME: use bulletpoints and terse descriptions.
- Link to v1: https://lore.kernel.org/r/20240420-d1_de2-v1-0-297efca674ba@jookia.org



--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 2,
    "change-id": "20240420-d1_de2-9be81d56bd85",
    "prefixes": [
      "RFC"
    ],
    "history": {
      "v1": [
        "20240420-d1_de2-v1-0-297efca674ba@jookia.org"
      ]
    }
  }
}