clk: Add helper for "fw:" clock name prefix
Clock functions like clk_register require clock parents to be addressed
by name. This doesn't work well when the clock's name is not known in
advance but instead supplied by a device tree.

A concrete example of this is on Allwinner boards: The 24MHz oscillator
is specified as a fixed-clock in the device tree then passed to the
clock controller unit by reference with the name "hosc".

This becomes a problem when drivers use the common clock framework
clocks to represent their clock trees: They can't specify a parent based
on the device tree node, only a global name.

To help solve this, introduce a helper that will resolve names with a
special prefix to the real clock's name. This helper will look for the
prefix "fw:", then resolve the name after it according to the device
tree's clock-names based on the clock's fwdev udevice.

For a hypothetical example, consider this device tree of a fixed
oscillator and a custom clock divider that divides a  controller:

        osc24m: osc24m-clk {
                compatible = "fixed-clock";
                clock-output-names = "osc24m";
                #clock-cells = <0>;
        };

        clock-divider {
                compatible = "my-divider";
                clocks = <&dcxo>;
                clock-names = "24mhz";
                #clock-cells = <1>;
                #reset-cells = <1>;
        };

Leveraging the common clock framework we can write the following code:

        clk_register_fixed_factor(dev, "12mhz", "fw:hosc", 0, 1, 2);
        clk_register_fixed_factor(dev,  "6mhz", "fw:hosc", 0, 1, 4);
        clk_register_fixed_factor(dev,  "3mhz", "fw:hosc", 0, 1, 8);

This will register a clock tree similar to this at runtime:

        osc24m (fixed-clock)
        ---- 12mhz (ccf_clk_fixed_factor)
        ---- 6mhz (ccf_clk_fixed_factor)
        ---- 3mhz (ccf_clk_fixed_factor)

Signed-off-by: John Watts <contact@jookia.org>