video: add support for MIPI DBI interface
Many simple RGB panels use a MIPI DBI interface over SPI, so this patch
adds support for this interface using an API compatible with Linux.

The MIPI DBI interface works similar to SPI using 8-bit words in the
format of: COMMAND DATA1 DATA2 ...
However the interface has two variations on this.

The first variation is implemented by this patch series: Using 9-bit
words, with the first bit signalling whether the word is a COMMAND or
DATA.

The second variation is not implemented by this series: A GPIO is held
at a value during an 8-bit word to indicate if it's a COMMAND or WORD.
Adding support for this wouldn't be too hard, but I don't have any
hardware to test it on.

Linux seems to not implement the specification correctly: It sets CS
inactive between words instead of keeping it active for the entire
transaction. I'm copying this behaviour to keep compatibility in porting
panels across from the kernel.
# Describe the purpose of this series. The information you put here
# will be used by the project maintainer to make a decision whether
# your patches should be reviewed, and in what priority order. Please be
# very detailed and link to any relevant discussions or sites that the
# maintainer can review to better understand your proposed changes. If you
# only have a single patch in your series, the contents of the cover
# letter will be appended to the "under-the-cut" portion of the patch.

# Lines starting with # will be removed from the cover letter. You can
# use them to add notes or reminders to yourself. If you want to use
# markdown headers in your cover letter, start the line with ">#".

# You can add trailers to the cover letter. Any email addresses found in
# these trailers will be added to the addresses specified/generated
# during the b4 send stage. You can also run "b4 prep --auto-to-cc" to
# auto-populate the To: and Cc: trailers based on the code being
# modified.

To: Jagan Teki <jagan@amarulasolutions.com>
To: Tom Rini <trini@konsulko.com>
To: Anatolij Gustschin <agust@denx.de>
Cc: u-boot@lists.denx.de
Signed-off-by: John Watts <contact@jookia.org>

---
Changes in v2:
- EDITME: describe what is new in this series revision.
- EDITME: use bulletpoints and terse descriptions.
- Link to v1: https://lore.kernel.org/r/20240614-mipi_dbi-v1-0-946eb82f8533@jookia.org



--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 2,
    "change-id": "20240601-mipi_dbi-aeda98f2f9d5",
    "prefixes": [],
    "history": {
      "v1": [
        "20240614-mipi_dbi-v1-0-946eb82f8533@jookia.org"
      ]
    }
  }
}